feat: remove direct envVars from Agent, enforce secrets-only env injection
