+++ apps/dashboard/src/entities/secret.ts
@@ -1,5 +1,18 @@
 import { z } from "zod";
 
+export const EnvVarSchema = z.object({
+  name: z
+    .string()
+    .min(1)
+    .regex(
+      /^[-._a-zA-Z0-9]+$/,
+      "A environment variable name must consist of alphanumeric characters, '-', '_' or '.'"
+    ),
+  value: z.string(),
+});
+
+export type EnvVar = z.infer<typeof EnvVarSchema>;
+
 export const SecretEnvVarSchema = z.object({
   name: z
     .string()
